Break out Checkbox, RadioButton and ImageButton to their own files
