+++ crates/egui/src/widgets/checkbox.rs
@@ -1,0 +1,136 @@
+use crate::*;
+
+// TODO(emilk): allow checkbox without a text label
+/// Boolean on/off control with text label.
+///
+/// Usually you'd use [`Ui::checkbox`] instead.
+///
+/// ```
+/// # egui::__run_test_ui(|ui| {
+/// # let mut my_bool = true;
+/// // These are equivalent:
+/// ui.checkbox(&mut my_bool, "Checked");
+/// ui.add(egui::Checkbox::new(&mut my_bool, "Checked"));
+/// # });
+/// ```
+#[must_use = "You should put this widget in an ui with `ui.add(widget);`"]
+pub struct Checkbox<'a> {
+    checked: &'a mut bool,
+    text: WidgetText,
+    indeterminate: bool,
+}
+
+impl<'a> Checkbox<'a> {
+    pub fn new(checked: &'a mut bool, text: impl Into<WidgetText>) -> Self {
+        Checkbox {
+            checked,
+            text: text.into(),
+            indeterminate: false,
+        }
+    }
+
+    pub fn without_text(checked: &'a mut bool) -> Self {
+        Self::new(checked, WidgetText::default())
+    }
+
+    /// Display an indeterminate state (neither checked nor unchecked)
+    ///
+    /// This only affects the checkbox's appearance. It will still toggle its boolean value when
+    /// clicked.
+    #[inline]
+    pub fn indeterminate(mut self, indeterminate: bool) -> Self {
+        self.indeterminate = indeterminate;
+        self
+    }
+}
+
+impl<'a> Widget for Checkbox<'a> {
+    fn ui(self, ui: &mut Ui) -> Response {
+        let Checkbox {
+            checked,
+            text,
+            indeterminate,
+        } = self;
+
+        let spacing = &ui.spacing();
+        let icon_width = spacing.icon_width;
+        let icon_spacing = spacing.icon_spacing;
+
+        let (galley, mut desired_size) = if text.is_empty() {
+            (None, vec2(icon_width, 0.0))
+        } else {
+            let total_extra = vec2(icon_width + icon_spacing, 0.0);
+
+            let wrap_width = ui.available_width() - total_extra.x;
+            let galley = text.into_galley(ui, None, wrap_width, TextStyle::Button);
+
+            let mut desired_size = total_extra + galley.size();
+            desired_size = desired_size.at_least(spacing.interact_size);
+
+            (Some(galley), desired_size)
+        };
+
+        desired_size = desired_size.at_least(Vec2::splat(spacing.interact_size.y));
+        desired_size.y = desired_size.y.max(icon_width);
+        let (rect, mut response) = ui.allocate_exact_size(desired_size, Sense::click());
+
+        if response.clicked() {
+            *checked = !*checked;
+            response.mark_changed();
+        }
+        response.widget_info(|| {
+            if indeterminate {
+                WidgetInfo::labeled(
+                    WidgetType::Checkbox,
+                    galley.as_ref().map_or("", |x| x.text()),
+                )
+            } else {
+                WidgetInfo::selected(
+                    WidgetType::Checkbox,
+                    *checked,
+                    galley.as_ref().map_or("", |x| x.text()),
+                )
+            }
+        });
+
+        if ui.is_rect_visible(rect) {
+            // let visuals = ui.style().interact_selectable(&response, *checked); // too colorful
+            let visuals = ui.style().interact(&response);
+            let (small_icon_rect, big_icon_rect) = ui.spacing().icon_rectangles(rect);
+            ui.painter().add(epaint::RectShape::new(
+                big_icon_rect.expand(visuals.expansion),
+                visuals.rounding,
+                visuals.bg_fill,
+                visuals.bg_stroke,
+            ));
+
+            if indeterminate {
+                // Horizontal line:
+                ui.painter().add(Shape::hline(
+                    small_icon_rect.x_range(),
+                    small_icon_rect.center().y,
+                    visuals.fg_stroke,
+                ));
+            } else if *checked {
+                // Check mark:
+                ui.painter().add(Shape::line(
+                    vec![
+                        pos2(small_icon_rect.left(), small_icon_rect.center().y),
+                        pos2(small_icon_rect.center().x, small_icon_rect.bottom()),
+                        pos2(small_icon_rect.right(), small_icon_rect.top()),
+                    ],
+                    visuals.fg_stroke,
+                ));
+            }
+            if let Some(galley) = galley {
+                let text_pos = pos2(
+                    rect.min.x + icon_width + icon_spacing,
+                    rect.center().y - 0.5 * galley.size().y,
+                );
+                ui.painter().galley(text_pos, galley, visuals.text_color());
+            }
+        }
+
+        response
+    }
+}
