+++ crates/egui/src/widgets/mod.rs
@@ -7,29 +7,37 @@
 use crate::*;
 
 mod button;
+mod checkbox;
 pub mod color_picker;
 pub(crate) mod drag_value;
 mod hyperlink;
 mod image;
+mod image_button;
 mod label;
 mod progress_bar;
+mod radio_button;
 mod selected_label;
 mod separator;
 mod slider;
 mod spinner;
 pub mod text_edit;
 
-pub use button::*;
-pub use drag_value::DragValue;
-pub use hyperlink::*;
-pub use image::{paint_texture_at, Image, ImageFit, ImageOptions, ImageSize, ImageSource};
-pub use label::*;
-pub use progress_bar::ProgressBar;
-pub use selected_label::SelectableLabel;
-pub use separator::Separator;
-pub use slider::*;
-pub use spinner::*;
-pub use text_edit::{TextBuffer, TextEdit};
+pub use self::{
+    button::Button,
+    checkbox::Checkbox,
+    drag_value::DragValue,
+    hyperlink::{Hyperlink, Link},
+    image::{paint_texture_at, Image, ImageFit, ImageOptions, ImageSize, ImageSource},
+    image_button::ImageButton,
+    label::Label,
+    progress_bar::ProgressBar,
+    radio_button::RadioButton,
+    selected_label::SelectableLabel,
+    separator::Separator,
+    slider::{Slider, SliderOrientation},
+    spinner::Spinner,
+    text_edit::{TextBuffer, TextEdit},
+};
 
 // ----------------------------------------------------------------------------
